requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyYAML==5.3.* 
uritemplate==3.0.*     